Fix FileExists error when a broken symlink exists
In the create_pack_script it used `os.path.exists` which returns false on broken symlinks. I changed it to `os.path.lexists`, which returns on True on broken symlinks. When the symlink was broken it thought there was no file and wanted to create a new symlink. This caused a FileExists error and the buildout crashed.